+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Create auth configmap containing MC name and admin group info
+
 ## [0.5.2] - 2023-07-04
 
 ### Changed